# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "RunPG API",
    "description": "API for the RunPG mobile game and watch app",
    "version": "1.0.0"
  },
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/inventory/{id}": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/inventory/{id}/quantity": {
      "put": {
        "tags": [
          "Inventory"
        ],
        "summary": "Update the quantity of an inventory",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "quantity": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User inventory updated"
          },
          "400": {
            "description": "id or quantity is not valid"
          },
          "404": {
            "description": "Inventory does not exist"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      }
    },
    "/spell/": {
      "get": {
        "tags": [
          "Spell"
        ],
        "summary": "Get all spells",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Got all spells"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      },
      "post": {
        "tags": [
          "Spell"
        ],
        "summary": "Create a new spell",
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/spell/name/{name}": {
      "get": {
        "tags": [
          "Spell"
        ],
        "summary": "Get a spell by name",
        "description": "",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Spell found"
          },
          "404": {
            "description": "Spell not found"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      }
    },
    "/spell/{id}": {
      "get": {
        "tags": [
          "Spell"
        ],
        "summary": "Get a spell by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Spell found"
          },
          "400": {
            "description": "spellId not valid"
          },
          "404": {
            "description": "Spell not found"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      }
    },
    "/guild/": {
      "get": {
        "tags": [
          "Guild"
        ],
        "summary": "Get all the guildes",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Guilds found"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      },
      "post": {
        "tags": [
          "Guild"
        ],
        "summary": "Create a guild",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "guild": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Guild created"
          },
          "400": {
            "description": "Guild body is not valid"
          },
          "409": {
            "description": "Guild already exists"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      }
    },
    "/guild/{guildId}": {
      "get": {
        "tags": [
          "Guild"
        ],
        "summary": "Get a guild by id",
        "description": "",
        "parameters": [
          {
            "name": "guildId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Guild found"
          },
          "400": {
            "description": "guildId is not valid"
          },
          "404": {
            "description": "Guild not found"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      },
      "put": {
        "tags": [
          "Guild"
        ],
        "summary": "Update a guild",
        "description": "",
        "parameters": [
          {
            "name": "guildId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "newGuildValues": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Guild updated"
          },
          "400": {
            "description": "Guild body or id is not valid"
          },
          "404": {
            "description": "Guild does not exists"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      }
    },
...
  }
}